Python/product_management_Assignment2/requirements.txt
fastapi
uvicorn[standard]
prisma
python-dotenv
pydantics